fix video parser
